removing base class and putting umbraco setup in UmbracoSupport
